+++ conduit-extra/Data/Conduit/Attoparsec.hs
@@ -53,11 +53,13 @@ instance Exception ParseError
 data Position = Position
     { posLine :: {-# UNPACK #-} !Int
     , posCol  :: {-# UNPACK #-} !Int
+    , posOffset :: {-# UNPACK #-} !Int
+    -- ^ @since 1.2.0
     }
     deriving (Eq, Ord)
 
 instance Show Position where
-    show (Position l c) = show l ++ ':' : show c
+    show (Position l c off) = show l ++ ':' : show c ++ " (" ++ show off ++ ")"
 
 data PositionRange = PositionRange
     { posRangeStart :: {-# UNPACK #-} !Position
@@ -88,10 +90,11 @@ instance AttoparsecInput B.ByteString where
     empty = B.empty
     isNull = B.null
     notEmpty = filter (not . B.null)
-    getLinesCols = B.foldl' f (Position 0 0)
+    getLinesCols = B.foldl' f (Position 0 0 0)
       where
-        f (Position l c) ch | ch == 10 = Position (l + 1) 0
-                            | otherwise = Position l (c + 1)
+        f (Position l c o) ch
+          | ch == 10 = Position (l + 1) 0 (o + 1)
+          | otherwise = Position l (c + 1) (o + 1)
     stripFromEnd b1 b2 = B.take (B.length b1 - B.length b2) b1
 
 instance AttoparsecInput T.Text where
@@ -100,10 +103,11 @@ instance AttoparsecInput T.Text where
     empty = T.empty
     isNull = T.null
     notEmpty = filter (not . T.null)
-    getLinesCols = T.foldl' f (Position 0 0)
+    getLinesCols = T.foldl' f (Position 0 0 0)
       where
-        f (Position l c) ch | ch == '\n' = Position (l + 1) 0
-                            | otherwise = Position l (c + 1)
+        f (Position l c o) ch
+          | ch == '\n' = Position (l + 1) 0 (o + 1)
+          | otherwise = Position l (c + 1) (o + 1)
     stripFromEnd (TI.Text arr1 off1 len1) (TI.Text _ _ len2) =
         TI.text arr1 off1 (len1 - len2)
 
@@ -114,14 +118,14 @@ instance AttoparsecInput T.Text where
 --
 -- Since 0.5.0
 sinkParser :: (AttoparsecInput a, MonadThrow m) => A.Parser a b -> Consumer a m b
-sinkParser = fmap snd . sinkParserPosErr (Position 1 1)
+sinkParser = fmap snd . sinkParserPosErr (Position 1 1 0)
 
 -- | Same as 'sinkParser', but we return an 'Either' type instead
 -- of raising an exception.
 --
 -- Since 1.1.5
 sinkParserEither :: (AttoparsecInput a, Monad m) => A.Parser a b -> Consumer a m (Either ParseError b)
-sinkParserEither = (fmap.fmap) snd . sinkParserPos (Position 1 1)
+sinkParserEither = (fmap.fmap) snd . sinkParserPos (Position 1 1 0)
 
 
 -- | Consume a stream of parsed tokens, returning both the token and
@@ -131,7 +135,7 @@ sinkParserEither = (fmap.fmap) snd . sinkParserPos (Position 1 1)
 -- Since 0.5.0
 conduitParser :: (AttoparsecInput a, MonadThrow m) => A.Parser a b -> Conduit a m (PositionRange, b)
 conduitParser parser =
-    conduit $ Position 1 1
+    conduit $ Position 1 1 0
        where
          conduit !pos = await >>= maybe (return ()) go
              where
@@ -158,7 +162,7 @@ conduitParserEither
     => A.Parser a b
     -> Conduit a m (Either ParseError (PositionRange, b))
 conduitParserEither parser =
-    conduit $ Position 1 1
+    conduit $ Position 1 1 0
   where
     conduit !pos = await >>= maybe (return ()) go
       where
@@ -232,10 +236,11 @@ sinkParserPos pos0 p = sink empty pos0 (parseA p)
                 pos' = addLinesCols prev pos
 
     addLinesCols :: AttoparsecInput a => a -> Position -> Position
-    addLinesCols x (Position lines cols) =
-        lines' `seq` cols' `seq` Position lines' cols'
+    addLinesCols x (Position lines cols off) =
+        lines' `seq` cols' `seq` off' `seq` Position lines' cols' off'
       where
-        Position dlines dcols = getLinesCols x
+        Position dlines dcols doff = getLinesCols x
         lines' = lines + dlines
         cols' = (if dlines > 0 then 1 else cols) + dcols
+        off' = off + doff
 {-# INLINE sinkParserPos #-}